ai-browser-agent/requirements.txt
browser-use
langchain-google-genai
langchain-openai
python-dotenv
playwright
